requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.15.0=pypi_0
addict=2.4.0=pypi_0
alsa-lib=1.2.3=h516909a_0
anyio=3.3.1=pypi_0
argon2-cffi=21.1.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
attrs=21.2.0=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bzip2=1.0.8=h7f98852_4
c-ares=1.17.2=h7f98852_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=4.2.2=pypi_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2024.2.2=py38h06a4308_0
cffi=1.14.6=pypi_0
charset-normalizer=2.0.4=pypi_0
clang=5.0=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
cycler=0.10.0=py_2
cython=0.29.28=py38h295c915_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2022.5.0=py38h06a4308_0
dbus=1.13.6=h48d8840_2
debugpy=1.4.3=pypi_0
decorator=5.1.0=pypi_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
entrypoints=0.3=pypi_0
etils=0.6.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.1=h9c3ff4c_0
ffmpeg=4.3.2=hca11adc_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
fsspec=2022.3.0=py38h06a4308_0
gast=0.3.3=pypi_0
geos=3.10.1=h9c3ff4c_1
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.3=h9c3ff4c_0
glib-tools=2.68.3=h9c3ff4c_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.56.3=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.32.0=pypi_0
gst-plugins-base=1.18.4=hf529b03_2
gstreamer=1.18.4=h76c114f_2
gviz-api=1.10.0=pypi_0
h5py=2.10.0=pypi_0
harfbuzz=2.8.2=h83ec7ef_0
hdf5=1.10.6=nompi_h6a2412b_1114
httplib2=0.20.4=pypi_0
icu=68.1=h58526e2_0
idna=3.2=pypi_0
image-classifiers=0.2.2=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
imgaug=0.4.0=pypi_0
importlib-resources=5.8.0=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.4.1=pypi_0
ipython=7.27.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jasper=1.900.1=h07fcdf6_1006
jbig=2.1=h7f98852_2003
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
json5=0.9.6=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=7.0.2=pypi_0
jupyter-core=4.7.1=pypi_0
jupyter-packaging=0.10.5=pypi_0
jupyter-server=1.11.0=pypi_0
jupyterlab=3.1.12=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-server=2.8.1=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kaleido=0.2.1=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py38h1fd1430_1
krb5=1.19.2=hcc1bbae_0
lame=3.100=h7f98852_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=11_linux64_mkl
libcblas=3.9.0=11_linux64_mkl
libclang=14.0.1=pypi_0
libcurl=7.78.0=h2574ce0_0
libdeflate=1.7=h7f98852_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.3=he6710b0_2
libgcc-ng=11.1.0=hc902ee8_8
libgfortran-ng=11.1.0=h69a702a_8
libgfortran5=11.1.0=h6c583b3_8
libglib=2.68.3=h3e27bee_0
libiconv=1.16=h516909a_0
liblapack=3.9.0=11_linux64_mkl
liblapacke=3.9.0=11_linux64_mkl
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.43.0=h812cca2_0
libogg=1.3.4=h7f98852_1
libopencv=4.5.2=py38hd24989e_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.3=hd57d9b9_0
libprotobuf=3.16.0=h780b84a_0
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=11.1.0=h56837e0_8
libtiff=4.3.0=hf544144_1
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
llvm-openmp=12.0.1=h4bd325d_1
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.3=py38h578d9bd_0
matplotlib-base=3.4.3=py38hf4fb855_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
nbclassic=0.3.1=pypi_0
nbclient=0.5.4=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
nettle=3.6=he412f7d_0
networkx=2.7.1=pyhd3eb1b0_0
notebook=6.4.4=pypi_0
nspr=4.30=h9c3ff4c_0
nss=3.69=hb5efdd6_0
numpy=1.19.5=pypi_0
numpy-base=1.20.3=py38h74d4b33_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py_0
open3d=0.13.0=pypi_0
opencv=4.5.2=py38h578d9bd_1
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.0=pypi_0
pandas=1.3.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pypi_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.2=pypi_0
pip=21.2.2=py38h06a4308_0
pixman=0.40.0=h36c2ea0_0
plotly=5.6.0=py_0
plotly-orca=1.3.1=1
plyfile=0.7.4=pypi_0
pot=0.8.1.0=py38h43a58ef_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.20=pypi_0
protobuf=3.17.3=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-opencv=4.5.2=py38hd0cf306_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.4=py38h6c62de6_0
pycparser=2.20=pypi_0
pygments=2.10.0=pypi_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38h578d9bd_7
pyqt-impl=5.12.3=py38h7400c14_7
pyqt5-sip=4.19.18=py38h709712a_7
pyqtchart=5.12=py38h7400c14_7
pyqtwebengine=5.12.1=py38h7400c14_7
pyrsistent=0.18.0=pypi_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.1=pypi_0
pywavelets=1.3.0=py38h7f8727e_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
qt=5.12.9=hda022c4_4
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
requests-unixsocket=0.2.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.19.2=py38h51133e4_0
scikit-learn=0.24.2=py38h1561384_1
scipy=1.4.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=52.0.0=py38h06a4308_0
shapely=1.8.0=py38h800f7b8_4
six=1.15.0=pypi_0
sniffio=1.2.0=pypi_0
sqlite=3.36.0=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
tenacity=8.0.1=py38h06a4308_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-profile=2.5.0=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.4.0=pypi_0
tensorflow-estimator=2.4.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
tensorflow-metadata=1.9.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
text-unidecode=1.3=pypi_0
tf-slim=1.1.0=pypi_0
threadpoolctl=2.2.0=pyh8a188c0_0
tifffile=2020.10.1=py38hdd07704_2
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
tomlkit=0.7.2=pypi_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py38h497a2fe_1
tqdm=4.62.0=pyhd8ed1ab_0
traitlets=5.1.0=pypi_0
trescope=0.0.1=pypi_0
trimesh=3.9.36=pyh6c4a22f_0
typeguard=2.13.3=pypi_0
typing-extensions=3.7.4.3=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.6=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.2.1=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.8.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.5.0=ha95c52a_0
